Publish initial version of tutorial
This initial commit has the first draft of the tutorial for optimizing
pandas data handling. You should expect grammar mistakes, and lots of
features and better code employment usage.

The release has a library called fasting_pandas for some simple modules
for benchmarking or creating and manipulating the datasets.

It is divided into 3 main jupyter notebooks that have all the
information and comments for the tutorial.

In future releases it is expected to have better module features for
benchmarking, better analytics, and some docs for additional
information.