Update tutorial and create benchmarking library
Fasting pandas can now be used as a benchmarking library. In the core.py
module you can find objects that extend the library by timing and
generating metadata from any method that is used in the pandas or
Dataframe objects. One of the nice features is that the metadata can be
saved into a sqlite3 database to store the results.

The main.py module shows an example of interacting with this library.
Benchmarks are generated by iterating through a sample dataframe of
different sizes and stored in a database. Finally, some graphs are
created to shows these benchmarks. The graphs need to be updated later.

One of the other features included with this commit is a new version of
the tutorial. Lots of context and additional information was added to
improve the user experience and didactical knowledge.

Heinz Stecher